Converting GUI to work with pyqwt
matplotlib is a bit of a slug, pyqwt is supposed to be faster.

This is a quick first shot, but it's already considerably faster. It
works, inasmuch as it draws the right lines. Picking doesn't happen, and
the colours are wrong, and it doesn't draw the fef, and and and...

Doing this is also revealing where the chart is too tightly coupled to
mpl, so some good will come out of having the two versions, in the form
of better code architecture.